[riscv64] Fill with the Machine Architecture ID Register `marchid` [ppc64] Added source comment

+++ riscv64/corefreq.h
@@ -33,7 +33,7 @@ typedef struct
 
 	struct
 	{
-		unsigned int		Revision;
+		unsigned long		Revision;
 
 		struct {
 		unsigned short int	CfgLock :  1-0,
